Functional test type 0: - New structure - Files replacing - First tests post type 0
