+++ test/functional/http/get/dapps.js
@@ -1,10 +1,10 @@
 'use strict';
 
-var node = require('./../node.js');
-var http = require('../common/httpCommunication.js');
-var ws = require('../common/wsCommunication.js');
-var clearDatabaseTable = require('../common/globalBefore').clearDatabaseTable;
-var modulesLoader = require('../common/initModule').modulesLoader;
+var node = require('../../../node.js');
+var http = require('../../../common/httpCommunication.js');
+var ws = require('../../../common/wsCommunication.js');
+var clearDatabaseTable = require('../../../common/globalBefore').clearDatabaseTable;
+var modulesLoader = require('../../../common/initModule').modulesLoader;
 
 function postTransaction (transaction, done) {
 	ws.call('postTransactions', { transaction: transaction }, done, true);
