+++ test/functional/http/get/delegates.js
@@ -1,11 +1,11 @@
 'use strict';
 
-var node = require('../node.js');
-var http = require('../common/httpCommunication.js');
-var sendLISK = require('../common/complexTransactions.js').sendLISK;
-var sendTransaction = require('../common/complexTransactions.js').sendTransaction;
-var modulesLoader = require('./../common/initModule.js').modulesLoader;
-var genesisDelegates = require('../genesisDelegates.json');
+var node = require('../../../node.js');
+var http = require('../../../common/httpCommunication.js');
+var sendLISK = require('../../../common/complexTransactions.js').sendLISK;
+var sendTransaction = require('../../../common/complexTransactions.js').sendTransaction;
+var modulesLoader = require('../../../common/initModule.js').modulesLoader;
+var genesisDelegates = require('../../../genesisDelegates.json');
 var testDelegate = genesisDelegates.delegates[0];
 
 //insert one extra delegate
