+++ test/functional/http/get/loader.js
@@ -1,7 +1,7 @@
 'use strict';
 
-var node = require('../node.js');
-var http = require('../common/httpCommunication.js');
+var node = require('../../../node.js');
+var http = require('../../../common/httpCommunication.js');
 
 describe('GET /api/loader/status/ping', function () {
