+++ test/functional/index.js
@@ -1,13 +1,13 @@
-require('./accounts');
-require('./blocks');
-require('./dapps');
-require('./delegates');
-require('./loader');
-require('./multisignatures');
-require('./peers');
-require('./transactions');
-require('./transport/transport');
-require('./transport/transport.blocks');
-require('./transport/transport.transactions');
-require('./transport/transport.transactions.stress');
-require('./transactions/0.tx');
+require('./http/get/accounts');
+require('./http/get/blocks');
+require('./http/get/dapps');
+require('./http/get/delegates');
+require('./http/get/loader');
+require('./http/get/multisignatures');
+require('./http/get/peers');
+require('./http/get/transactions');
+require('./ws/transport');
+require('./ws/transport.blocks');
+require('./ws/transport.transactions');
+require('./ws/transport.transactions.stress');
+require('./http/post/0.tx');
